pyproject.toml
[tool.poetry]
name = "weather-monitoring"
version = "0.1.0"
description = ""
authors = ["anvitha31303 <2000031303@kluniversity.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
